+++ src/savefile.h
@@ -93,6 +93,7 @@ int rd_quests_1(void);
 int rd_quests_2(void);
 int rd_quests_3(void);
 int rd_artifacts(void);
+int rd_player_1(void);
 int rd_player(void);
 int rd_ignore(void);
 int rd_misc(void);